benchmark_scripts/requirements.txt
absl-py==2.0.0
attrs==23.1.0
cloudpickle==2.2.1
decorator==5.1.1
exceptiongroup==1.1.3
iniconfig==2.0.0
ml-dtypes==0.3.1
numpy==1.26.0
packaging==23.2
pluggy==1.3.0
psutil==5.9.5
pytest==7.4.2
scipy==1.11.3
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
xgboost==2.0.0
